Pull apart OrderCompletionReset service
